requirements.txt
pyserial
shutil
subprocess
esptool
